Add TLS Health Check (#363)
Add new gRPC health check that serves as a readiness check for the scaler pods. This health check ensures that the TLS certificates can be successfully read from the volume and loaded into memory.

This PR also includes a refactor around the TLS services to better enable the integration of the health checks by leveraging the new keyed services in .NET 8.